requirements.txt
flask
flask-mysqldb
web3
python-dotenv
werkzeug
